refactor(injection): extract shared useInjectionSource hook
Deduplicate the identical effect/ref/cleanup logic from
useCommentInjection and usePrCommentInjection into a
generic useInjectionSource<T> hook.
