+++ src/renderer/hooks/usePrCommentInjection.ts
@@ -1,7 +1,7 @@
-import { useEffect, useRef, useSyncExternalStore } from 'react';
+import { useCallback, useSyncExternalStore } from 'react';
 import { selectedPrCommentsStore } from '../lib/selectedPrCommentsStore';
-import { pendingInjectionManager } from '../lib/PendingInjectionManager';
 import { formatPrCommentsForAgent } from '../lib/formatPrCommentsForAgent';
+import { useInjectionSource } from './useInjectionSource';
 
 const INJECTION_SOURCE = 'pr-comments';
 
@@ -14,42 +14,12 @@ const getSnapshot = () => selectedPrCommentsStore.getSnapshot();
  */
 export function usePrCommentInjection() {
   const selected = useSyncExternalStore(subscribe, getSnapshot);
-  const hasPendingRef = useRef(false);
 
-  useEffect(() => {
-    if (selected.length === 0) {
-      if (hasPendingRef.current) {
-        pendingInjectionManager.clear(INJECTION_SOURCE);
-        hasPendingRef.current = false;
-      }
-      return () => {
-        if (hasPendingRef.current) {
-          pendingInjectionManager.clear(INJECTION_SOURCE);
-          hasPendingRef.current = false;
-        }
-      };
-    }
-
-    const formatted = formatPrCommentsForAgent(selected);
-    if (formatted) {
-      pendingInjectionManager.setPending(formatted, INJECTION_SOURCE);
-      hasPendingRef.current = true;
-    }
-
-    return () => {
-      if (hasPendingRef.current) {
-        pendingInjectionManager.clear(INJECTION_SOURCE);
-        hasPendingRef.current = false;
-      }
-    };
-  }, [selected]);
-
-  useEffect(() => {
-    return pendingInjectionManager.onInjectionUsed(() => {
-      selectedPrCommentsStore.clear();
-      hasPendingRef.current = false;
-    });
+  const onConsumed = useCallback(() => {
+    selectedPrCommentsStore.clear();
   }, []);
+
+  useInjectionSource(INJECTION_SOURCE, selected, formatPrCommentsForAgent, onConsumed);
 }
 
 /**
